cert/requirements.txt
transformers==4.12.5
sentencepiece
protobuf
wandb
tqdm
datasets
tensorboard
fairseq
fairscale
zstandard
openpyxl
matplotlib
pandas>=1.1.2
torchvision>=0.7.0
seaborn>=0.11.2
pyyaml
ipdb
numpy
cython
autopep8
docformatter
beautifulsoup4
lxml
redbaron
func-timeout